+++ spec/clustering/KMeansModel_spec.lua
@@ -2,7 +2,7 @@ local hasSparkSession, _ = pcall(require, 'stuart-sql.SparkSession')
 local KMeansModel = require 'stuart-ml.clustering.KMeansModel'
 local registerAsserts = require 'registerAsserts'
 local stuart = require 'stuart'
---local Vectors = require 'stuart-ml.linalg.Vectors'
+local Vectors = require 'stuart-ml.linalg.Vectors'
 
 registerAsserts(assert)
 
@@ -21,9 +21,9 @@ describe('clustering.KMeansModel', function()
     local sc = stuart.NewContext()
     local model = KMeansModel.load(sc, 'spec-fixtures/model4')
     assert.equal(2, #model.clusterCenters)
---    assert.same({1,2,3}, model.clusterCenters:toArray())
---    assert.same({5,6,7}, model.clusterCenters:toArray())
---    assert.equal(1, model:predict(Vectors.dense({5,6,7})))
+    assert.same({5,6,7}, model.clusterCenters[1]:toArray())
+    assert.same({1,2,3}, model.clusterCenters[2]:toArray())
+    assert.equal(1, model:predict(Vectors.dense({5,6,7})))
   end)
 
 end)